10-Azure-SDK-Course/02-Manage-Resource-group/requirements.txt
azure.mgmt.resource
azure.identity
python-dotenv